+++ src/services/ClientsService.ts
@@ -1,18 +1,21 @@
-import { TablesUpdate } from "@/database/database";
-import supabase from "@/database/supabase";
-import { ClientRecord } from "@/types";
+import { TablesUpdate } from "@/database/database"
+import supabase from "@/database/supabase"
+import { ClientRecord } from "@/types"
 
 export const deleteClient = async (id: string | number) => {
-  const { error } = await supabase.from("customer").delete().eq("id", id);
+    const {error} = await supabase.
+    from('customer').
+    delete().
+    eq('id', id)
 
-  if (error) {
-    console.log(error);
-    return false;
-  }
-  return true;
-};
+    if(error){
+        console.log(error)
+        return false
+    }
+    return true
+}
 
-export const fetchAllClients = async () => {
+export const getAllClients = async () => {
     const { data, error } = await supabase.from('customer').select('*')
 
     if(error){
@@ -31,5 +34,15 @@ export const updateClient = async (client: any) => {
     }
     return true
 }
+
+export const getClientById = async (id: string | number) => {
+    const { data, error } = await supabase.from('customer').select('*').eq('id', id).single()
+
+    if(error){
+        console.log(error)
+        return null
+    }
+    return data
+    }
